recycler view animation

+++ app/src/main/java/com/samsofa/expiredhelper/ui/MainActivity.java
@@ -5,7 +5,6 @@ import android.graphics.Canvas;
 import android.os.Bundle;
 import android.util.TypedValue;
 import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.View.OnClickListener;
@@ -13,6 +12,7 @@ import android.widget.RelativeLayout;
 import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import androidx.core.content.ContextCompat;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
@@ -29,6 +29,7 @@ import com.samsofa.expiredhelper.models.Item;
 import com.samsofa.expiredhelper.viewModels.ItemViewModel;
 import it.xabaras.android.recyclerview.swipedecorator.RecyclerViewSwipeDecorator;
 import java.util.List;
+import java.util.Objects;
 
 
 public class MainActivity extends AppCompatActivity implements RecyclerViewClickListener {
@@ -39,73 +40,41 @@ public class MainActivity extends AppCompatActivity implements RecyclerViewClick
 
   private ItemAdapter adapter;
 
+  Toolbar toolbar;
+
   private RecyclerView mRecyclerView;
-  ItemTouchHelper.SimpleCallback simpleCallback = new SimpleCallback(0, ItemTouchHelper.LEFT) {
-    @Override
-    public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull ViewHolder viewHolder,
-        @NonNull ViewHolder target) {
-      return false;
-    }
 
-    @Override
-    public void onChildDraw(@NonNull Canvas c, @NonNull RecyclerView recyclerView,
-        @NonNull ViewHolder viewHolder, float dX, float dY, int actionState,
-        boolean isCurrentlyActive) {
-      new RecyclerViewSwipeDecorator.Builder(c, mRecyclerView, viewHolder, dX, dY, actionState,
-          isCurrentlyActive)
-          .addBackgroundColor(ContextCompat.getColor(MainActivity.this, R.color.red))
-          .addActionIcon(R.drawable.ic_delete)
-          .addSwipeLeftLabel("Delete")
-          .setSwipeLeftLabelColor(R.color.black)
-          .setSwipeLeftLabelTextSize(TypedValue.COMPLEX_UNIT_SP, 16.0f)
-          .create()
-          .decorate();
-
-      super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-    }
-
-    @Override
-    public void onSwiped(@NonNull ViewHolder viewHolder, int direction) {
-
-      if (direction == ItemTouchHelper.LEFT) {
-        itemViewModel.delete(adapter.getItemAt(viewHolder.getAdapterPosition()));
-
-        Toast.makeText(MainActivity.this,
-            "item at position deleted: " + viewHolder.getAdapterPosition(), Toast.LENGTH_SHORT)
-            .show();
-      }
-
-
-    }
-  };
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_main);
 
+    toolbar = findViewById(R.id.toolbar);
+    setSupportActionBar(toolbar);
+
     emptyView = findViewById(R.id.empty_view);
-
-    adapter = new ItemAdapter(this);
-    mRecyclerView = findViewById(R.id.main_recylerView);
-    mRecyclerView
-        .setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
-    mRecyclerView.setAdapter(adapter);
-
+    emptyView.setVisibility(View.VISIBLE);
     itemViewModel = new ViewModelProvider(this,
         ViewModelProvider.AndroidViewModelFactory.getInstance(this.getApplication()))
         .get(ItemViewModel.class);
+
+    adapter = new ItemAdapter(this, MainActivity.this);
+    mRecyclerView = findViewById(R.id.main_recylerView);
+    mRecyclerView
+        .setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
+
+    mRecyclerView.setHasFixedSize(true);
 
     itemViewModel.getAllItems().observe(this, new Observer<List<Item>>() {
       @Override
       public void onChanged(List<Item> items) {
         adapter.submitList(items);
-
-        if (adapter.getCurrentList().size() > 0) {
-          emptyView.setVisibility(View.GONE);
-        }
+        dealEmptyViewVisibility(items);
       }
     });
+
+    mRecyclerView.setAdapter(adapter);
 
     FloatingActionButton fab = findViewById(R.id.fab);
     fab.setOnClickListener(new OnClickListener() {
@@ -116,37 +85,44 @@ public class MainActivity extends AppCompatActivity implements RecyclerViewClick
       }
     });
 
-    ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleCallback);
-    itemTouchHelper.attachToRecyclerView(mRecyclerView);
+    new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
 
-//    new ItemTouchHelper(new SimpleCallback(0, ItemTouchHelper.LEFT) {
-//      @Override
-//      public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull ViewHolder viewHolder,
-//          @NonNull ViewHolder target) {
-//        return false;
-//      }
-//
-//      @Override
-//      public void onSwiped(@NonNull ViewHolder viewHolder, int direction) {
-//
-//        itemViewModel.delete(adapter.getItemsAt(viewHolder.getAdapterPosition()));
-//        Toast.makeText(MainActivity.this, "Item deleted " + viewHolder.getAdapterPosition(), Toast.LENGTH_SHORT).show();
-//      }
-////
-////      @Override
-////      public void onChildDraw(@NonNull Canvas c, @NonNull RecyclerView recyclerView,
-////          @NonNull ViewHolder viewHolder, float dX, float dY, int actionState,
-////          boolean isCurrentlyActive) {
-////        new RecyclerViewSwipeDecorator.Builder(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive)
-////            .addSwipeLeftActionIcon(R.drawable.ic_delete)
-////            .addBackgroundColor(ContextCompat.getColor(MainActivity.this, R.color.red))
-////            .addSwipeLeftLabel("Delete")
-////            .create()
-////            .decorate();
-////        super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-////      }
-//    }).attachToRecyclerView(mRecyclerView);
-//
+      @Override
+      public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull ViewHolder viewHolder,
+          @NonNull ViewHolder target) {
+        return false;
+      }
+
+      @Override
+      public void onChildDraw(@NonNull Canvas c, @NonNull RecyclerView recyclerView,
+          @NonNull ViewHolder viewHolder, float dX, float dY, int actionState,
+          boolean isCurrentlyActive) {
+        new RecyclerViewSwipeDecorator.Builder(c, mRecyclerView, viewHolder, dX, dY, actionState,
+            isCurrentlyActive)
+            .addBackgroundColor(ContextCompat.getColor(MainActivity.this, R.color.gray))
+            .addActionIcon(R.drawable.ic_delete)
+            .addSwipeLeftLabel("Delete")
+            .setSwipeLeftLabelColor(R.color.red)
+            .setSwipeLeftActionIconTint(R.color.red)
+            //
+            //.setSwipeLeftLabelTypeface()
+            //      .setSwipeLeftLabelTextSize(TypedValue.COMPLEX_UNIT_SP, 16.0f)
+            .create()
+            .decorate();
+
+        super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
+      }
+
+      @Override
+      public void onSwiped(@NonNull ViewHolder viewHolder, int direction) {
+        if (direction == ItemTouchHelper.LEFT) {
+          itemViewModel.delete(adapter.getItemAt(viewHolder.getAdapterPosition()));
+          Toast.makeText(MainActivity.this,
+              "item at position deleted: " + viewHolder.getAdapterPosition(), Toast.LENGTH_SHORT)
+              .show();
+        }
+      }
+    }).attachToRecyclerView(mRecyclerView);
 
   }
 
@@ -166,18 +142,33 @@ public class MainActivity extends AppCompatActivity implements RecyclerViewClick
 
   @Override
   public boolean onCreateOptionsMenu(Menu menu) {
-    MenuInflater menuInflater = getMenuInflater();
-    menuInflater.inflate(R.menu.main_menu, menu);
+    getMenuInflater().inflate(R.menu.main_menu, menu);
     return true;
   }
 
   @Override
   public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-    switch (item.getItemId()) {
-      case R.id.delete_all_items:
+    if (item.getItemId() == R.id.action_delete_all_items) {
+      if (Objects.requireNonNull(itemViewModel.getAllItems().getValue()).size() > 0) {
         itemViewModel.deleteAllItems();
         Toast.makeText(this, "all Items deleted", Toast.LENGTH_SHORT).show();
+        return true;
+      } else {
+        Toast.makeText(this, "no Items found", Toast.LENGTH_SHORT).show();
+      }
     }
     return super.onOptionsItemSelected(item);
   }
+
+  private void dealEmptyViewVisibility(List<Item> items) {
+    if (items.size() > 0) {
+      mRecyclerView.setVisibility(View.VISIBLE);
+      emptyView.setVisibility(View.GONE);
+    } else {
+      mRecyclerView.setVisibility(View.GONE);
+      emptyView.setVisibility(View.VISIBLE);
+    }
+  }
+
+
 }
